Fix navigate after post a job and f5 error
